Fix builds on Linux
The builds were failing because of the OS mismatch in the Gemfile.lock.
GitHub's  macOS 11 runner is still private, so I can't just match the
workflow to my machine. The error message from the build says that this
bundle incantation might do the trick.

However, bundler won't be available until after the ruby/setup-ruby
action runs, so I have to disable the `bundle install` caching from that
step and do it manually instead.