Randomize tail-based sampling
We can get away with the cheaper `rand` operation here because we're
tacitly assuming that the processor will be passed the entire trace in
tail mode. We only do the SHA nonsense in head mode so that the "coin
flip" (that isn't really a coin flip) lands the same way for each event
in a trace.